Add RemoveChars Op halfway
This tries to add a RemoveChars op with options until I realized
I should probably be using 3 subtypes of remove char ops. This is
because Symbols would be a lot easier to do with a regex.

Also had a bunch of difficulty getting these to work with Custom
settings, while also letting users set it from code.